chore: rm ts-loader, enhance generated type declarations
