+++ package.json
@@ -36,7 +36,6 @@
     "@patternfly/react-table": "^4.93.1",
     "@rhoas/app-services-ui-shared": "^0.16.0",
     "ajv": "^8.8.2",
-    "dts-loader": "^0.1.7",
     "immer": "^9.0.15",
     "monaco-editor": "^0.33.0",
     "monaco-editor-webpack-plugin": "^7.0.1",
