+++ src/@kaoto/index.d.ts
@@ -24,17 +24,6 @@ export interface IDeployment {
   type: string;
 }
 
-export interface ISettings {
-  description?: string;
-  // e.g. 'KameletBinding'
-  dsl: string;
-  icon?: string;
-  // name of integration or deployment
-  name: string;
-  // Cluster namespace
-  namespace: string;
-}
-
 export interface IIntegration {
   metadata: IIntegrationMetadata;
   params: IIntegrationParams[];
@@ -58,6 +47,45 @@ export interface IIntegrationParams {
   path?: boolean;
   type?: string;
   value?: {};
+}
+
+export interface ISettings {
+  description?: string;
+  // e.g. 'KameletBinding'
+  dsl: string;
+  icon?: string;
+  // name of integration or deployment
+  name: string;
+  // Cluster namespace
+  namespace: string;
+}
+
+/**
+ * The API for a typical Step Extension
+ * The following are methods that are exposed to a Step Extension.
+ */
+export interface IStepExtensionApi {
+  getDeployment: (name: string, namespace?: string) => Promise<string | unknown>;
+  getIntegrationSource: (
+    integration: IIntegration,
+    dsl: string,
+    namespace?: string
+  ) => Promise<string | unknown>;
+  getStep: () => IStepProps;
+  notifyKaoto: (title: string, body?: string, variant?: string) => void;
+  onKaotoButtonClicked: (view: IViewProps) => void;
+  saveConfig: (newValues: { [s: string]: unknown } | ArrayLike<unknown>) => void;
+  startDeployment: (
+    integration: any,
+    name: string,
+    namespace?: string
+  ) => Promise<string | unknown>;
+  step: IStepProps;
+  stepParams: { [s: string]: unknown };
+  stepInitialValues: { [p: string]: any };
+  stopDeployment: (name: string, namespace?: string) => void;
+  updateStep: (step: IStepProps) => void;
+  updateStepParams: (newValues: { [s: string]: unknown } | ArrayLike<unknown>) => void;
 }
 
 export interface IStepProps {
