+++ src/api/apiService.ts
@@ -1,5 +1,5 @@
-import { IIntegration, IStepProps } from '../types';
 import request from './request';
+import { IIntegration, IStepProps } from '@kaoto';
 
 const apiVersion = '/v1';
 
@@ -241,7 +241,7 @@ export async function fetchViews(data: IStepProps[], namespace?: string) {
  */
 export async function startDeployment(integrationSource: string, name: string, namespace?: string) {
   try {
-    const params = (namespace?`?namespace=${namespace}`:'');
+    const params = namespace ? `?namespace=${namespace}` : '';
     const resp = await request.post({
       endpoint: `${apiVersion}/deployments/${name.toLowerCase()}${params}`,
       contentType: 'text/yaml',
